updating files

+++ General_C/maths/math.c
@@ -11,10 +11,10 @@ int mathssub(int n1, int n2)
 }
 long long int maths_mul(int n1, int n2)
 {
-    return n1*n2;
+    return (long long int)n1*n2;
 }
 float maths_div(int n1, int n2)
 {
-    return n1/n2;
+    return (float)n1/n2;
 }
